backend/module1/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9
pdfplumber==0.11.0
PyPDF2==3.0.1
google-generativeai==0.8.0
pydantic==2.9.0
python-dotenv==1.0.1
aiofiles==24.1.0
